+++ src/track/track.service.ts
@@ -1,3 +1,5 @@
+import { AlbumService } from './../album/album.service';
+import { ArtistService } from './../artist/artist.service';
 import { FavouriteService } from './../favourite/favourite.service';
 import { ERRORS } from './../types/Error';
 import { v4 } from 'uuid';
@@ -6,6 +8,7 @@ import {
   NotFoundException,
   forwardRef,
   Inject,
+  UnprocessableEntityException,
 } from '@nestjs/common';
 import { CreateTrackDto } from './dto/create-track.dto';
 import { UpdateTrackDto } from './dto/update-track.dto';
@@ -14,13 +17,39 @@ import { TrackEntity } from './entities/track.entity';
 @Injectable()
 export class TrackService {
   constructor(
+    @Inject(forwardRef(() => AlbumService))
+    private readonly albumService: AlbumService,
+    @Inject(forwardRef(() => ArtistService))
+    private readonly artistService: ArtistService,
     @Inject(forwardRef(() => FavouriteService))
     private readonly favouriteService: FavouriteService,
   ) {}
 
   private static tracks: TrackEntity[] = [];
 
-  create(createTrackDto: CreateTrackDto): Promise<TrackEntity> {
+  async create(createTrackDto: CreateTrackDto): Promise<TrackEntity> {
+    if (createTrackDto.artistId) {
+      try {
+        await this.artistService.findOne(createTrackDto.artistId);
+      } catch (error) {
+        if (error.status === 404)
+          throw new UnprocessableEntityException(ERRORS.NOT_FOUND);
+
+        throw error;
+      }
+    }
+
+    if (createTrackDto.albumId) {
+      try {
+        await this.albumService.findOne(createTrackDto.albumId);
+      } catch (error) {
+        if (error.status === 404)
+          throw new UnprocessableEntityException(ERRORS.NOT_FOUND);
+
+        throw error;
+      }
+    }
+
     const track = new TrackEntity({
       id: v4(),
       ...createTrackDto,
@@ -58,7 +87,12 @@ export class TrackService {
 
     if (!track) throw new NotFoundException(ERRORS.NOT_FOUND);
 
-    await this.favouriteService.remove({ id, type: 'tracks' });
+    const { tracks } = await this.favouriteService.findAll();
+    const hasInFavourites = tracks.some(({ id: trackId }) => trackId === id);
+
+    if (hasInFavourites) {
+      await this.favouriteService.remove({ id, type: 'tracks' });
+    }
 
     TrackService.tracks = TrackService.tracks.filter(
       ({ id: trackId }) => trackId !== id,
